requirements.txt
Flask>=2.2.0
pymongo>=4.0.0
python-dotenv>=0.20.0
Flask-Login>=0.6.0
Werkzeug>=2.2.0
bcrypt>=4.0.0
Authlib>=1.0.0
pyodbc
WeasyPrint>=56.0
APScheduler>=3.9.0